2023-05-17 Fred Gleason <fredg@paravelsystems.com> 	* Fixed a regression in the WebAPI that caused imports to bypass the 	maximum file size limitation set the 'System Settings' in rdadmin(1).

+++ lib/rdformpost.h
@@ -37,7 +37,7 @@ class RDFormPost
   enum Encoding {UrlEncoded=0,MultipartEncoded=1,AutoEncoded=2};
   enum Error {ErrorOk=0,ErrorNotPost=1,ErrorNoTempDir=2,ErrorMalformedData=3,
 	      ErrorPostTooLarge=4,ErrorInternal=5,ErrorNotInitialized=6};
-  RDFormPost(RDFormPost::Encoding encoding,unsigned maxsize=0,
+  RDFormPost(RDFormPost::Encoding encoding,int64_t maxsize=0,
 	     bool auto_delete=true);
   ~RDFormPost();
   RDFormPost::Error error() const;
@@ -76,7 +76,7 @@ class RDFormPost
   QMap<QString,bool> post_filenames;
   RDTempDirectory *post_tempdir;
   bool post_auto_delete;
-  unsigned post_content_length;
+  int64_t post_content_length;
   QString post_content_type;
   char *post_data;
   QString post_separator;
